fix critical bugs: PI truncation, type mismatches, typos, and unchecked I/O (Phase 1)
- Fix `stopInitTime = false` → `0.0` in pathFollower.cpp
- Replace truncated PI (3.1415926) with M_PI from <cmath> across 9 files
- Fix `amend_cmake` → `ament_cmake` typo in arise_slam_mid360/package.xml
- Add NULL guards for fopen in visualizationTools.cpp to prevent fprintf crash
- Rename arize_slam.launch.py → arise_slam.launch.py and update 6 references
- Fix signed/unsigned mismatches: 36 int→size_t for .size() across 8 files

+++ src/autonomy/base_autonomy/sensor_scan_generation/src/sensorScanGeneration.cpp
@@ -59,12 +59,12 @@ void laserCloudAndOdometryHandler(const nav_msgs::msg::Odometry::ConstSharedPtr 
   transformToMap.setRotation(tf2::Quaternion(odometryIn.pose.pose.orientation.x, odometryIn.pose.pose.orientation.y,
                                             odometryIn.pose.pose.orientation.z, odometryIn.pose.pose.orientation.w));
 
-  int laserCloudInNum = laserCloudIn->points.size();
+  size_t laserCloudInNum = laserCloudIn->points.size();
 
   pcl::PointXYZ p1;
   tf2::Vector3 vec;
 
-  for (int i = 0; i < laserCloudInNum; i++)
+  for (size_t i = 0; i < laserCloudInNum; i++)
   {
     p1 = laserCloudIn->points[i];
     vec.setX(p1.x);
